헤어샵, 메뉴, 디자이너 테스트 본인인증 로직 추가

+++ src/main/java/com/prgms/kokoahairshop/menu/controller/MenuController.java
@@ -1,9 +1,12 @@
 package com.prgms.kokoahairshop.menu.controller;
 
+import com.prgms.kokoahairshop.hairshop.entity.Hairshop;
 import com.prgms.kokoahairshop.menu.dto.CreateMenuRequest;
 import com.prgms.kokoahairshop.menu.dto.MenuResponse;
 import com.prgms.kokoahairshop.menu.dto.ModifyMenuRequest;
+import com.prgms.kokoahairshop.menu.entity.Menu;
 import com.prgms.kokoahairshop.menu.service.MenuService;
+import com.prgms.kokoahairshop.user.entity.User;
 import java.net.URI;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +14,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -29,7 +33,11 @@ public class MenuController {
     private final MenuService menuService;
 
     @PostMapping
-    public ResponseEntity<Long> insert(@Valid @RequestBody CreateMenuRequest createMenuRequest) {
+    public ResponseEntity<Long> insert(@Valid @RequestBody CreateMenuRequest createMenuRequest, @AuthenticationPrincipal User user) {
+        Hairshop hairshop = menuService.findHairshopById(createMenuRequest.getHairshopId());
+        if (!hairshop.getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("본인의 헤어샵 메뉴만 생성할 수 있습니다.");
+        }
         MenuResponse insert = menuService.insert(createMenuRequest);
         return ResponseEntity.created(URI.create("/menu/" + insert.getId())).body(insert.getId());
     }
@@ -52,13 +60,21 @@ public class MenuController {
     }
 
     @PatchMapping
-    public ResponseEntity<Object> modify(@Valid @RequestBody ModifyMenuRequest modifyMenuRequest) {
+    public ResponseEntity<Object> modify(@Valid @RequestBody ModifyMenuRequest modifyMenuRequest, @AuthenticationPrincipal User user) {
+        Hairshop hairshop = menuService.findHairshopById(modifyMenuRequest.getHairshopId());
+        if (!hairshop.getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("본인의 헤어샵 메뉴만 수정할 수 있습니다.");
+        }
         menuService.update(modifyMenuRequest);
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deleteById(@PathVariable Long id) {
+    public ResponseEntity<Object> deleteById(@PathVariable Long id, @AuthenticationPrincipal User user) {
+        Menu menu = menuService.findMenuById(id);
+        if (!menu.getHairshop().getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("본인의 헤어샵 메뉴만 삭제할 수 있습니다.");
+        }
         menuService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
